rename notifications promo sheet
