+++ src/navigation/Routes.android.js
@@ -30,7 +30,7 @@ import SettingsSheet, {
 import ShowcaseSheet from '../screens/ShowcaseSheet';
 import SpeedUpAndCancelSheet from '../screens/SpeedUpAndCancelSheet';
 import SwapsPromoSheet from '../screens/SwapsPromoSheet';
-import NotificationsPromoSheet from '../screens/NotificationsPromoSheet';
+import NotificationsPromoSheet from '../screens/notifications-promo';
 import TransactionConfirmationScreen from '../screens/TransactionConfirmationScreen';
 import WalletConnectApprovalSheet from '../screens/WalletConnectApprovalSheet';
 import WalletConnectRedirectSheet from '../screens/WalletConnectRedirectSheet';
